+++ stacker/blueprints/rds/base.py
@@ -123,6 +123,14 @@ class BaseRDS(Blueprint):
                 "default": "false",
                 "allowed_values": ["true", "false"]
             },
+            "StorageType": {
+                "type": "String",
+                "description": "Storage type for RDS instance. Defaults to "
+                               "standard unless IOPS is set, then it "
+                               "defaults to io1",
+                "default": "default",
+                "allowed_values": ["default", "standard", "gp2", "io1"]
+            },
             "AllocatedStorage": {
                 "type": "Number",
                 "description": "Space, in GB, to allocate to RDS instance. If "
@@ -166,6 +174,11 @@ class BaseRDS(Blueprint):
                 "max_length": "63",
                 "allowed_pattern": "[a-zA-Z][a-zA-Z0-9-]*",
                 "default": self.name},
+            "DBSnapshotIdentifier": {
+                "type": "String",
+                "description": "The snapshot you want the db restored from.",
+                "default": "",
+            },
             "EngineVersion": {
                 "type": "String",
                 "description": "Database engine version for the RDS Instance.",
@@ -225,11 +238,14 @@ class BaseRDS(Blueprint):
                 Condition("HasInternalZoneName"),
                 Condition("HasInternalHostname")))
         t.add_condition(
-            "ProvisionedIOPS",
+            "HasProvisionedIOPS",
             Not(Equals(Ref("IOPS"), "0")))
         t.add_condition(
-            "NoProvisionedIOPS",
-            Equals(Ref("IOPS"), "0"))
+            "HasStorageType",
+            Not(Equals(Ref("StorageType"), "default")))
+        t.add_condition(
+            "HasDBSnapshotIdentifier",
+            Not(Equals(Ref("DBSnapshotIdentifier"), "")))
 
     def create_subnet_group(self):
         t = self.template
@@ -285,14 +301,13 @@ class BaseRDS(Blueprint):
         t.add_resource(
             DBInstance(
                 DBINSTANCE,
-                StorageType=If("NoProvisionedIOPS",
-                               Ref("AWS::NoValue"),
-                               "io1"),
-                Iops=If("NoProvisionedIOPS",
-                        Ref("AWS::NoValue"),
-                        Ref("IOPS")),
-                **self.get_common_attrs()
-                ))
+                StorageType=If("HasStorageType",
+                               Ref("StorageType"),
+                               Ref("AWS::NoValue")),
+                Iops=If("HasProvisionedIOPS",
+                        Ref("IOPS"),
+                        Ref("AWS::NoValue")),
+                **self.get_common_attrs()))
 
     def create_dns_records(self):
         t = self.template
@@ -392,6 +407,11 @@ class MasterInstance(BaseRDS):
             "DBName": Ref("DatabaseName"),
             "DBInstanceClass": Ref("InstanceType"),
             "DBInstanceIdentifier": Ref("DBInstanceIdentifier"),
+            "DBSnapshotIdentifier": If(
+                "HasDBSnapshotIdentifier",
+                Ref("DBSnapshotIdentifier"),
+                Ref("AWS::NoValue"),
+            ),
             "DBParameterGroupName": Ref("ParameterGroup"),
             "DBSubnetGroupName": Ref(SUBNET_GROUP),
             "Engine": self.engine() or Ref("Engine"),
@@ -427,6 +447,7 @@ class ReadReplica(BaseRDS):
             "AutoMinorVersionUpgrade": Ref("AutoMinorVersionUpgrade"),
             "DBInstanceClass": Ref("InstanceType"),
             "DBInstanceIdentifier": Ref("DBInstanceIdentifier"),
+            "DBParameterGroupName": Ref("ParameterGroup"),
             "Engine": self.engine() or Ref("Engine"),
             "EngineVersion": Ref("EngineVersion"),
             "OptionGroupName": Ref("OptionGroup"),